move chatops helper methods to modules included by Command
