+++ spec/lib/chat_ops/commands/list_responder_command_spec.rb
@@ -1,5 +1,6 @@
 RSpec.describe ChatOps::Commands::ListResponderCommand do
   include ChatOpsCommandHelper
+  include ChatOps::MentionHelpers
 
   describe 'regex' do
     it_should_match_commands <<-EOL
@@ -17,14 +18,14 @@ RSpec.describe ChatOps::Commands::ListResponderCommand do
       incident.responders << user1
       incident.save
 
-      expect(process("incident responders")).to return_response_matching /#{ChatOps.prevent_highlights(user1.name)}/
+      expect(process("incident responders")).to return_response_matching /#{prevent_highlights(user1.name)}/
     end
 
     it "should list multiple responders" do
       incident.responders << user1
       incident.responders << user2
       incident.save
-      expect(process("incident responders")).to return_response_matching(/#{ChatOps.prevent_highlights(user1.name)}/, /#{ChatOps.prevent_highlights(user2.name)}/, /#{incident.incident_id}/)
+      expect(process("incident responders")).to return_response_matching(/#{prevent_highlights(user1.name)}/, /#{prevent_highlights(user2.name)}/, /#{incident.incident_id}/)
     end
   end
 end
