+++ spec/lib/chat_ops/message_helpers_spec.rb
@@ -1,0 +1,3 @@
+RSpec.describe ChatOps::MessageHelpers do
+  let(:helper) { Class.new { extend ChatOps::MessageHelpers } }
+end
